requirements.txt
flask~=3.0.0
aiohttp~=3.8.6
requests~=2.31.0
pydantic~=2.5.1
duckdb~=0.9.2
Faker~=20.0.3
gunicorn==20.0.4
google-cloud-storage